+++ epdlib/constants.py
@@ -8,8 +8,13 @@ from pathlib import Path
 
 absolute_path = Path(__file__).resolve().parent
 
+MODES = {'1': '1 bit', 'L': '8 bit', 'RGB': 'Red, Green, Blue'}
 
 SCREEN_ROTATIONS = [0, 90, -90, 180, 270]
+
+COLORS_7 = ['RED', 'ORANGE', 'YELLOW', 'GREEN', 'BLUE', 'WHITE', 'BLACK']
+
+CLEAR_COLOR = 0xFF
 
 LAYOUT_DEFAULTS = {'type': None,
                    'image': None,